+++ controllers/json/AttributeController.php
@@ -6,13 +6,17 @@ use Yii;
 use app\classes\JsonController;
 use app\models\Attribute;
 use app\exceptions\FormValidationException;
+use yii\web\ForbiddenHttpException;
 use yii\web\HttpException;
 
 class AttributeController extends JsonController
 {
     public function actionList()
     {
-
+        if (!\Yii::$app->user->can('attribute_list')) {
+            throw new ForbiddenHttpException('Access denied');
+        }
+        
         return
             Attribute::find()
                 ->select(['id', 'name'])
@@ -23,6 +27,9 @@ class AttributeController extends JsonController
 
     public function actionRead()
     {
+        if (!\Yii::$app->user->can('attribute_list')) {
+            throw new ForbiddenHttpException('Access denied');
+        }
 
         return
             Attribute::find()
@@ -34,6 +41,10 @@ class AttributeController extends JsonController
 
     public function actionGet()
     {
+        if (!\Yii::$app->user->can('attribute_edit')) {
+            throw new ForbiddenHttpException('Access denied');
+        }
+        
         $item = Attribute::findOne($this->request['id']);
         if ($item === null) {
             throw new HttpException(404, 'Attribute не найден');
@@ -44,10 +55,17 @@ class AttributeController extends JsonController
 
     public function actionSave()
     {
-
         if (isset($this->request['id'])) {
+            if (!\Yii::$app->user->can('attribute_edit')) {
+                throw new ForbiddenHttpException('Access denied');
+            }
+            
             $item = $this->getAttributeOr404($this->request['id']);
         } else {
+            if (!\Yii::$app->user->can('attribute_create')) {
+                throw new ForbiddenHttpException('Access denied');
+            }
+            
             $item = Attribute::create();
         }
 
@@ -68,6 +86,10 @@ class AttributeController extends JsonController
 
     public function actionDelete()
     {
+        if (!\Yii::$app->user->can('attribute_delete')) {
+            throw new ForbiddenHttpException('Access denied');
+        }
+        
         $item = Attribute::findOne($this->request['id']);
         $item->delete();
     }
