+++ controllers/json/AttributeGroupController.php
@@ -6,13 +6,17 @@ use Yii;
 use app\classes\JsonController;
 use app\models\AttributeGroup;
 use app\exceptions\FormValidationException;
+use yii\web\ForbiddenHttpException;
 use yii\web\HttpException;
 
 class AttributeGroupController extends JsonController
 {
     public function actionList()
     {
-
+        if (!\Yii::$app->user->can('attribute_group_list')) {
+            throw new ForbiddenHttpException('Access denied');
+        }
+        
         return
             AttributeGroup::find()
                 ->select(['id', 'name'])
@@ -23,7 +27,10 @@ class AttributeGroupController extends JsonController
 
     public function actionRead()
     {
-
+        if (!\Yii::$app->user->can('attribute_group_list')) {
+            throw new ForbiddenHttpException('Access denied');
+        }
+        
         return
             AttributeGroup::find()
                 ->select(['id', 'name', 'attributeslist_ids','note'])
@@ -34,6 +41,10 @@ class AttributeGroupController extends JsonController
 
     public function actionGet()
     {
+        if (!\Yii::$app->user->can('attribute_group_edit')) {
+            throw new ForbiddenHttpException('Access denied');
+        }
+        
         $item = AttributeGroup::findOne($this->request['id']);
         if ($item === null) {
             throw new HttpException(404, 'AttributeGroup не найден');
@@ -44,10 +55,17 @@ class AttributeGroupController extends JsonController
 
     public function actionSave()
     {
-
         if (isset($this->request['id'])) {
+            if (!\Yii::$app->user->can('attribute_group_edit')) {
+                throw new ForbiddenHttpException('Access denied');
+            }
+            
             $item = $this->getAttributeGroupOr404($this->request['id']);
         } else {
+            if (!\Yii::$app->user->can('attribute_group_create')) {
+                throw new ForbiddenHttpException('Access denied');
+            }
+            
             $item = AttributeGroup::create();
         }
 
@@ -70,6 +88,10 @@ class AttributeGroupController extends JsonController
 
     public function actionDelete()
     {
+        if (!\Yii::$app->user->can('attribute_group_delete')) {
+            throw new ForbiddenHttpException('Access denied');
+        }
+        
         $item = AttributeGroup::findOne($this->request['id']);
         $item->delete();
     }
